3-1 Project : MathIsland
마지막이네 고생했다
최종빌드.

+++ Assets/Scripts/Stage1_Puz7.cs
@@ -6,19 +6,24 @@ public class Stage1_Puz7 : MonoBehaviour
 {
     [SerializeField] PuzzleSlots Slot;
     [SerializeField] Gate gate;
+    AudioSource Audio;
+    bool invoked;
 
 	// Use this for initialization
 	void Start ()
     {
-		
-	}
+        Audio = GetComponent<AudioSource>();
+        invoked = false;
+    }
 	
 	// Update is called once per frame
 	void Update ()
     {
-		if(Slot.Satisfied)
+		if(Slot.Satisfied && !invoked)
         {
+            invoked = true;
             gate.Open(0.3f);
+            Audio.Play();
             PlayerPrefs.DeleteKey("BLOCK 3");
         }
 	}
